+++ CHANGES.md
@@ -12,6 +12,7 @@
 * Add `moduleHeatshield.required_resources`
 * Add binding for transmitter modules
 * Add binding for lighting modules
+* Add binding for generator modules
 
 ## 0.5.4.0 -> 0.5.4.1